+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/bitbucketserver/BitBucketTypes.kt
@@ -1,25 +1,29 @@
 package com.github.jensim.megamanipulator.actions.vcs.bitbucketserver
 
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+
+@Serializable
 data class BitBucketMessage(
     // val context: String?
     val message: String,
     val exceptionName: String?
 )
 
+@Serializable
 data class BitBucketPage<T>(
-    val start: Int?,
-    val size: Int?,
-    val limit: Int?,
-    val isLastPage: Boolean?,
-    val values: List<T>?,
-    val message: String?,
+    val start: Int? = null,
+    val size: Int? = null,
+    val limit: Int? = null,
+    val isLastPage: Boolean? = null,
+    val values: List<T>? = null,
+    val message: String? = null,
 )
 
-data class BitBucketPullRequest(
-    val id: Long? = null,
-    val version: Int? = null,
+@Serializable
+data class BitBucketPullRequestRequest(
     val title: String,
-    val description: String?,
+    val description: String? = null,
     val state: String = "OPEN",
     val open: Boolean = true,
     val closed: Boolean = false,
@@ -30,27 +34,49 @@ data class BitBucketPullRequest(
     val reviewers: List<BitBucketParticipant>,
 )
 
+@Serializable
+data class BitBucketPullRequest(
+    val id: Long? = null,
+    val version: Int? = null,
+    val title: String,
+    val description: String? = null,
+    val state: String? = null,
+    val open: Boolean? = null,
+    val closed: Boolean? = null,
+    val locked: Boolean? = null,
+    val fromRef: BitBucketBranchRef,
+    val toRef: BitBucketBranchRef,
+    @Transient
+    val author: BitBucketParticipant? = null,
+    val reviewers: List<BitBucketParticipant>,
+)
+
+@Serializable
 data class BitBucketParticipant(
     val user: BitBucketUser
 )
 
+@Serializable
 data class BitBucketUser(
     val name: String,
     val displayName: String? = null,
     val emailAddress: String? = null,
 )
 
+@Serializable
 data class BitBucketBranchRef(
     val id: String,
     val repository: BitBucketRepo,
 )
 
+@Serializable
 data class BitBucketDefaultBranch(
     val id: String, // "refs/heads/main",
     val displayId: String, // "main",
     val type: String, // "BRANCH"
 )
 
+@Serializable
 data class BitBucketRepo(
     val id: Long? = null,
     val slug: String,
@@ -59,6 +85,7 @@ data class BitBucketRepo(
     val links: BitBucketRepoLinks? = null,
 )
 
+@Serializable
 data class BitBucketForkRequest(
     val slug: String,
     val name: String? = null,
@@ -66,24 +93,29 @@ data class BitBucketForkRequest(
     val defaultBranch: String? = null,
 )
 
+@Serializable
 data class BitBucketProjectRequest(
-    val key: String
+    val key: String,
 )
 
+@Serializable
 data class BitBucketRepoLinks(
     val clone: List<BitBucketCloneLink>,
     val self: List<BitBucketPlainLink>,
 )
 
+@Serializable
 data class BitBucketCloneLink(
     val href: String,
     val name: String,
 )
 
+@Serializable
 data class BitBucketPlainLink(
-    val href: String
+    val href: String,
 )
 
+@Serializable
 data class BitBucketProject(
-    val key: String
+    val key: String,
 )